% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_management.R
\name{add_config_entry}
\alias{add_config_entry}
\title{Adds an entry to a DB config.yml file}
\usage{
add_config_entry(
  name,
  dbname,
  dbuser = dbname,
  file = "conf/config.yml",
  where = c("development", "production"),
  ...
)
}
\arguments{
\item{name}{Entry name in the config file}

\item{dbname}{Database name}

\item{dbuser}{User name. Default is database name.}

\item{where}{Development or production. If developments, adds an entry block to both default and development.}

\item{\dots}{Further arguments passed to \code{db_entry_list}}
}
\description{
Manage your DB config file with ease. Adds one or more config blocks,
in default, development, or production (or default+development at the same time since
they are both for the dev2 postgres server). Prompts to enter the password.
The password is not yet encrypted - working on it.
}
\details{
A shinto config file (as of 2022) is a YAML (typically \code{conf/config.yml}) contains postgres
database connection details including passwords. An entry looks like:

\preformatted{

  Eindhoven:
    dbname: "wbm_eindhoven"
    dbhost: "127.0.0.1"
    dbport: 2222
    dbuser: "wbm_eindhoven@postgres-dev2"
    dbpassword: "<<PASSWORD>>"
}

Details:
\itemize{
\item \code{dbname} is the name of the Postgres database
\item \code{dbhost} is the address; here it is localhost used in the port forwarding scenario (see below)
\item \code{dbport} is the port used in local forwarding (varies with user, see below)
\item \code{dbpassword} is the password as stored in 1PassWord. Not yet encrypted but watch this space!
}

A connection can then be made with \code{shintobag::shinto_db_connection("Eindhoven")}.

The port used in local forwarding is read from the environment variable \code{SHINTO_DEV2_LOCAL_PORT}.
Or, a menu appears where you can enter it (but it is not remembered).

A config file consists of three blocks: default, development and production. Which block
is used in by \code{\link[shintobag:shinto_db_connection]{shintobag::shinto_db_connection()}}) depends on the environment variable \code{R_CONFIG_ACTIVE}.
If it is not set, \code{default} is used, and otherwise the value of the variable. On the
shinto rsconnect servers, \code{devapp.shintolabs.net} has the value \code{development} for
this environment variable, and \code{app.shintolabs.net} a value of \code{production}. More
could follow in the future.

It is a convention for shinto postgres databases to have equal username and database names (for use in applications).
However this is not always the case!
}
\examples{
\dontrun{
add_config_entry("Testdatabase", "data_test", file = "test/config.yml", where = "development")
}
}
\seealso{
\code{\link[shintobag:shinto_db_connection]{shintobag::shinto_db_connection()}}
}
