% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browser_info_dependencies.R
\name{browser_info_dependencies}
\alias{browser_info_dependencies}
\title{Browser/navigator info dependencies}
\usage{
browser_info_dependencies()
}
\description{
In the UI, place the command \code{shintoshiny::browser_info_dependencies()}.
}
