% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_management.R
\name{check_config}
\alias{check_config}
\title{Check a config for presence of one or more database connections}
\usage{
check_config(name, where, file = "conf/config.yml")
}
\description{
Check a config for presence of one or more database connections
}
