% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/password_encrypt.R
\name{config_is_encrypted}
\alias{config_is_encrypted}
\title{Check if a config file is encrypted}
\usage{
config_is_encrypted(
  file,
  password_names = "dbpassword",
  secret = Sys.getenv("SHINTO_PASS_SECRET")
)
}
\description{
Checks whether a \code{config.yml} file is completely or partially encrypted.
Returns TRUE when all \code{dbpassword} (or other \code{password_names}) entries are already encrypted,
FALSE when one or more are not yet encrypted (with a message)
}
\examples{
\dontrun{
Sys.setenv(SHINTO_PASS_SECRET = "somesecret")
config_is_encrypted("conf/config.yml")
}
}
\seealso{
\code{\link{encrypt_config_file}}
}
