% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_management.R
\name{db_entry_list}
\alias{db_entry_list}
\title{Makes a config entry list}
\usage{
db_entry_list(
  name,
  dbname,
  dbuser = dbname,
  infra = c("dev2", "p2"),
  local = FALSE,
  password = NULL,
  encrypt = FALSE
)
}
\description{
Used by \code{\link[=add_config_entry]{add_config_entry()}}. See there for help.
}
\examples{
library(glue)
db_entry_list("Waalre","wbm_waalre", infra = "dev2", local = FALSE)  
}
