% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/password_encrypt.R
\name{encrypt_config_file}
\alias{encrypt_config_file}
\title{Encrypts a config file}
\usage{
encrypt_config_file(
  file,
  secret = Sys.getenv("SHINTO_PASS_SECRET"),
  suffix = ".encrypted",
  password_names = "dbpassword"
)
}
\arguments{
\item{file}{A YAML file with password entries}

\item{secret}{Secret used in symmetric encryption (see \code{\link{encrypt}})}

\item{suffix}{Suffix for new file name}

\item{password_names}{Entries to look for with passwords.}
}
\description{
Finds password fields in a YAML config file, and encrypts them.
The results are always written to a new file, with suffix ".encrypted" (or provided
\code{suffix} argument).
}
\examples{
\dontrun{
Sys.setenv(SHINTO_PASS_SECRET = "mysecret")
encrypt_config_file("config.yml")
}
}
