% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idletimeout.R
\name{idle_timeout}
\alias{idle_timeout}
\title{Stop application after idle timeout}
\usage{
idle_timeout(time = 10, units = c("minutes", "seconds"))
}
\arguments{
\item{time}{Time after which 'stopApp' is called}

\item{units}{Either 'minutes' or 'seconds' of time.}
}
\description{
Stop application after idle timeout
}
\details{
jQuery implementation
}
