% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_shinto_r_packages.R
\name{install_shinto_rpackages}
\alias{install_shinto_rpackages}
\title{Install (all) Shinto Labs R packages}
\usage{
install_shinto_rpackages(
  what = c("shintobag", "shintoshiny", "shintoui", "shinyfilterset", "shintodashboard",
    "shintousers", "shinysavedom", "shintopostcode")
)
}
\arguments{
\item{what}{Which packages to install. Defaults to most common ones.}
}
\description{
Install (all) Shinto Labs R packages
}
