% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_deploy_project.R
\name{make_deploy_project}
\alias{make_deploy_project}
\title{Make a project to deploy / update an application}
\usage{
make_deploy_project(
  name,
  path = "c:/repos_deploy",
  ignore_dirs = c("scripts", "stubs", "test", "tests", "backup", "docs", "rsconnect",
    ".git", ".Rproj.user"),
  ignore_files = c(".gitignore", ".Rhistory", "[.]Rproj$")
)
}
\arguments{
\item{name}{Name of the project, normally: client_appname ("eindhoven_kamerverhuur", "riec_rro").}

\item{path}{Where to make the new project, the project will be created in a subdirectory,
given by its \code{name}.}

\item{ignore_dirs}{Directories to ignore, these contain content that should **not** be uploaded
to Rstudio Connect.}

\item{ignore_files}{Files to ignore (usually these are files in the root, otherwise use
\code{ignore_dirs}).}
}
\description{
Opens a copy of this project including only relevant files,
from which you can deploy the application to Shinto Labs' Rstudio Connect server.
}
\examples{
\dontrun{

# Opens a new project "nederweert_buitengebied", a copy of the current repository
# from which you can deploy to rsconnect (via the deploy button), after removing
# unwanted files (data for other clients, etc.), and setting client-specific
# parameters (for example, in "this_version.yml")
shintoconnect::deploy("nederweert_buitengebied")
}
}
