% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_dependencies_info.R
\name{package_dependencies_info}
\alias{package_dependencies_info}
\title{Info for all required packages for this application}
\usage{
package_dependencies_info()
}
\description{
Info for all required packages for this application
}
