% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_helpers.R
\name{check_for_version}
\alias{check_for_version}
\alias{load_modules}
\title{Check for package version}
\usage{
check_for_version(package, version)

load_modules()
}
\description{
Check for package version

Load modules in a Shinto Shiny app
}
