% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsconnect_deployments.R
\name{log_rsconnect_deployments}
\alias{log_rsconnect_deployments}
\alias{connect_db_rsconnect_deployments}
\title{Write deployment info to rsconnect_deployments database}
\usage{
log_rsconnect_deployments(con, appname, environment, userid)

connect_db_rsconnect_deployments(config_file)
}
\description{
Write deployment info to rsconnect_deployments database
}
