% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R, R/shinto_db_connection.R
\name{shinto_db_connection}
\alias{shinto_db_connection}
\title{Connect to shinto devpostgres02}
\usage{
shinto_db_connection(
  what = c("CBS", "BRKdata", "BAGdata", "shintoanalytics"),
  file = "conf/config.yml"
)

shinto_db_connection(
  what = c("CBS", "BRKdata", "BAGdata", "shintoanalytics"),
  file = "conf/config.yml"
)
}
\description{
Connect to shinto devpostgres02

Connect to shinto devpostgres02
}
