% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{shinto_write_user_login}
\alias{shinto_write_user_login}
\title{Schrijf user login data
Voor gebruik in een shiny app. Zet eerst 'nav' met shintobrowserinfo, vind user (via lokale get_user, bv.),
en geef naam van applicatie door.}
\usage{
shinto_write_user_login(
  user = "unknown",
  application = "package_test",
  version = "",
  db = NULL,
  nav = NULL,
  session = shiny::getDefaultReactiveDomain()
)
}
\description{
Schrijf user login data
Voor gebruik in een shiny app. Zet eerst 'nav' met shintobrowserinfo, vind user (via lokale get_user, bv.),
en geef naam van applicatie door.
}
