% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shintoshiny_dependencies.R
\name{shintoshiny_dependencies}
\alias{shintoshiny_dependencies}
\title{Browser/navigator info dependencies}
\usage{
shintoshiny_dependencies()
}
\description{
In the UI, place the command \code{shintoshiny::browser_info_dependencies()}.
}
