% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsconnect_deployments.R
\name{log_rsconnect_deployments}
\alias{log_rsconnect_deployments}
\alias{connect_db_rsconnect_deployments}
\title{Write deployment info to rsconnect_deployments database}
\usage{
log_rsconnect_deployments(
  con,
  appname,
  environment,
  userid,
  shintoconnect_manifest = NULL,
  version = NULL
)

connect_db_rsconnect_deployments(config_file)
}
\arguments{
\item{con}{Database connection to \code{rsconnect_deployments}}

\item{appname}{app name on posit connect}

\item{environment}{Environment field in db (filled elsewhere)}

\item{userid}{userid}
}
\description{
Write deployment info to rsconnect_deployments database
}
