% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_inputs.R
\name{select_input}
\alias{select_input}
\alias{picker_input}
\alias{material_switch}
\title{Common shiny inputs with Shinto defaults}
\usage{
select_input(...)

picker_input(..., dropdown_align_right = TRUE, dropup_auto = TRUE)

material_switch(..., status = "primary", inline = TRUE)
}
\description{
Common shiny inputs with Shinto defaults
}
