% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StratumUser.R
\name{StratumUser}
\alias{StratumUser}
\title{A Stratum User R6 Object}
\description{
A Stratum User R6 Object

A Stratum User R6 Object
}
\details{
Alleen gebruikt voor shinyproxy deployments met AD authenticatie.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{StratumUser$new()}}
\item \href{#method-roles}{\code{StratumUser$roles()}}
\item \href{#method-roles_table}{\code{StratumUser$roles_table()}}
\item \href{#method-has_role}{\code{StratumUser$has_role()}}
\item \href{#method-dump}{\code{StratumUser$dump()}}
\item \href{#method-clone}{\code{StratumUser$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StratumUser$new(
  secret = NULL,
  jwt_preset = NULL,
  session = NULL,
  required_role = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-roles"></a>}}
\if{latex}{\out{\hypertarget{method-roles}{}}}
\subsection{Method \code{roles()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StratumUser$roles()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-roles_table"></a>}}
\if{latex}{\out{\hypertarget{method-roles_table}{}}}
\subsection{Method \code{roles_table()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StratumUser$roles_table()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-has_role"></a>}}
\if{latex}{\out{\hypertarget{method-has_role}{}}}
\subsection{Method \code{has_role()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StratumUser$has_role(role, local = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dump"></a>}}
\if{latex}{\out{\hypertarget{method-dump}{}}}
\subsection{Method \code{dump()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StratumUser$dump()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StratumUser$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
